backend/requirements.txt
# backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
groq==0.11.0
pymupdf==1.24.11
sentence-transformers==3.2.1
faiss-cpu==1.9.0
numpy==1.26.4
pydantic==2.9.2
python-multipart==0.0.12
supabase==2.9.1
